fix(go-core): change errorMessage.StatusCode from int to uint16 to match HTTP status code range
HTTP status codes are defined in the range 0-999 (RFC 9110) but the code was
using int with manual bounds checking. Change StatusCode to uint16 throughout
the error handling path to enforce the valid range at the type level.

Changes:
- errors.go: StatusCode field and createErrorMessage parameter changed from
  int to uint16; parseError() uses strconv.ParseUint with 16-bit limit
  instead of Atoi + manual bounds check
- index.go: remove redundant uint16() c

+++ golang/index.go
@@ -749,7 +749,7 @@ func dispatcherAsync(res fullRequest, chanWrite *safeChannelWriter) {
 			b.WriteByte(0)
 			b.WriteByte(5)
 			b.WriteString("error")
-			statusCode743 := uint16(parsedError.StatusCode)
+			statusCode743 := parsedError.StatusCode
 			b.WriteByte(byte(statusCode743 >> 8))
 			b.WriteByte(byte(statusCode743))
 
@@ -860,7 +860,7 @@ func dispatcherAsync(res fullRequest, chanWrite *safeChannelWriter) {
 					b.WriteByte(0)
 					b.WriteByte(5)
 					b.WriteString("error")
-					statusCode853 := uint16(parsedError.StatusCode)
+					statusCode853 := parsedError.StatusCode
 					b.WriteByte(byte(statusCode853 >> 8))
 					b.WriteByte(byte(statusCode853))
 
@@ -1783,7 +1783,7 @@ func (client CycleTLS) Do(URL string, options Options, Method string) (Response,
 	if err != nil {
 		parsedError := parseError(err)
 		return Response{
-			Status: parsedError.StatusCode,
+			Status: int(parsedError.StatusCode),
 			Body:   parsedError.ErrorMsg + " -> " + err.Error(),
 		}, nil
 	}
